third_party/blink/web_tests/platform/generic/fast/events/frame-click-clear-focus-expected.txt

Make sure that clicking on an inactive frame clears existing element focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

===> Making a frame with a focused element inactive by focus() for another frame.
PASS doc.activeElement is doc.body
===> Making a frame with a focused element inactive by mouse click for another frame.
PASS doc.activeElement is doc.body
===> Making the frame active again clicking by a non-focusable element.
PASS doc.activeElement is doc.body
PASS successfullyParsed is true

TEST COMPLETE

 
